+++ FindPanel/OgreTextFindComponents/NSTableView/OgreTableColumnAdapter.m
@@ -90,7 +90,7 @@
 	NSLog(@"  -numberOfChildrenInSelection: of %@", [self className]);
 #endif
     if ([_tableColumn isKindOfClass:[OgreTableColumn class]]) {
-        int count = [[_tableColumn tableView] numberOfSelectedRows];
+        NSInteger count = [[_tableColumn tableView] numberOfSelectedRows];
         if (inSelection && (count > 0)) return count;
         
         return [[_tableColumn tableView] numberOfRows];
@@ -105,7 +105,7 @@
 	NSLog(@"  -childAtIndex: of %@", [self className]);
 #endif
     OgreTableCellAdapter *tableCellAdapter;
-    unsigned            rowIndex;
+    NSUInteger            rowIndex;
     
     if (!inSelection) {
         rowIndex = index;
@@ -144,7 +144,7 @@
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -componentEnumeratorInSelection: of %@", [self className]);
 #endif
-    int count = [[_tableColumn tableView] numberOfSelectedRows];
+    NSInteger count = [[_tableColumn tableView] numberOfSelectedRows];
     
     OgreTextFindComponentEnumerator *enumerator;
     if ([self isReversed]) {
