+++ FindPanel/OgreTextFindComponents/NSTextView/OgreTextViewUndoer.m
@@ -52,8 +52,8 @@
 	NSTextStorage       *textStorage = [aTarget textStorage];
     NSRange             aRange, newRange;
     NSAttributedString  *aString;
-    unsigned            i;
-    OgreTextViewUndoer    *redoArray = [[OgreTextViewUndoer alloc] initWithCapacity:_count];
+    NSUInteger          i;
+    OgreTextViewUndoer  *redoArray = [[OgreTextViewUndoer alloc] initWithCapacity:_count];
     
     [textStorage beginEditing];
